Avance1PIA/requirements.txt
requests==2.26.0
openpyxl==3.0.7
virustotal_api==1.1.11
Pillow==8.4.0
beautifulsoup4==4.10.0